requirements.txt
amqp==5.2.0
asgiref==3.5.0
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0.3
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.1
kombu==5.3.6
MarkupSafe==2.1.1
numpy==1.26.4
oauthlib==3.2.0
prompt-toolkit==3.0.43
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2022.1
redis==5.0.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.1.0
wcwidth==0.2.13
